# cursillo/swagger/dist/schema/swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"This is the Cursillo Swagger documentation.",
    "version":"1.0.0",
    "title":"Cursillo Swagger "
  },
  "host":"localhost",
  "basePath":":8080/api/",
...
  "schemes":[
    "http"
  ],
  "paths":{
...
    "/talks/{talkId}/people": {
      "get":{
        "tags":[
          "talk"
        ],
        "summary":"Return people that have had talk with Id talkId.",
        "description":"",
        "operationId":"getPeopleByTalkId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"talkId",
            "description":"Talk Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
...
    "/people": {
      "get":{
        "tags":[
          "person"
        ],
        "summary":"Gets all people.",
        "description":"",
        "operationId":"Get People",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "post":{
        "tags":[
          "person"
        ],
        "summary":"Creates a new person.",
        "description":"",
        "operationId":"createPerson",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new person",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Person"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/people/{personId}": {
      "get":{
        "tags":[
          "person"
        ],
        "summary":"Return person with Id personId.",
        "description":"",
        "operationId":"getPersonbyId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"personId",
            "description":"Person Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "put":{
        "tags":[
          "person"
        ],
        "summary":"Updates an existing Person.",
        "description":"",
        "operationId":"updatePerson",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"personId",
            "description":"Id of person being updated.",
            "required":true
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new person.",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Person"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "delete":{
        "tags":[
          "person"
        ],
        "summary":"Delete Person with Id personId.",
        "description":"",
        "operationId":"deletePersonById",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"personId",
            "description":"Person Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/people/{personId}/talkLinks": {
      "get":{
        "tags":[
          "person"
        ],
        "summary":"Return person with Id personId.",
        "description":"",
        "operationId":"getTalkLinksByPersonId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"personId",
            "description":"Person Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
...
  },
  "securityDefinitions":{
    "petstore_auth":{
      "type":"oauth2",
      "authorizationUrl":"http://petstore.swagger.io/oauth/dialog",
      "flow":"implicit",
      "scopes":{
        "write:pets":"modify pets in your account",
        "read:pets":"read your pets"
      }
    },
    "api_key":{
      "type":"apiKey",
      "name":"api_key",
      "in":"header"
    }
  },
  "definitions":{
...
    "Person": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "required": true
        },
        "firstName": {
          "type": "string",
          "required": true
        },
        "lastName": {
          "type": "string",
          "required": true
        },
        "email": {
          "type": "string",
          "required": true
        },
        "homePhone": {
          "type": "string"
        },
        "workPhone": {
          "type": "string"
        },
        "street": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "zip": {
          "type": "number"
        },
        "birthdate": {
          "type": "string"
        },
        "occupation": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "isTeamMember": {
          "type": "boolean",
          "required": true,
          "default": false
        },
        "dateStarted": {
          "type": "string"
        },
        "tagName": {
          "type": "string"
        },
        "parishId": {
          "type": "integer",
          "format": "int64",
          "required": true
        }
      }
    },
...
  }
}